Refactor Agent interface
