--- crates/goose-cli/src/agents/agent.rs
@@ -1,43 +1,0 @@
-// use anyhow::Result;
-use async_trait::async_trait;
-// use futures::stream::BoxStream;
-use goose::{
-    agents::Agent, providers::base::Provider, providers::base::ProviderUsage, systems::System,
-};
-use tokio::sync::Mutex;
-
-pub struct GooseAgent {
-    systems: Vec<Box<dyn System>>,
-    provider: Box<dyn Provider>,
-    provider_usage: Mutex<Vec<ProviderUsage>>,
-}
-
-#[allow(dead_code)]
-impl GooseAgent {
-    pub fn new(provider: Box<dyn Provider>) -> Self {
-        Self {
-            systems: Vec::new(),
-            provider,
-            provider_usage: Mutex::new(Vec::new()),
-        }
-    }
-}
-
-#[async_trait]
-impl Agent for GooseAgent {
-    fn add_system(&mut self, system: Box<dyn System>) {
-        self.systems.push(system);
-    }
-
-    fn get_systems(&self) -> &Vec<Box<dyn System>> {
-        &self.systems
-    }
-
-    fn get_provider(&self) -> &Box<dyn Provider> {
-        &self.provider
-    }
-
-    fn get_provider_usage(&self) -> &Mutex<Vec<ProviderUsage>> {
-        &self.provider_usage
-    }
-}
