+++ crates/goose-cli/src/agents/mod.rs
@@ -1,4 +1,2 @@
-pub mod agent;
-
 #[cfg(test)]
 pub mod mock_agent;
