+++ crates/goose-cli/src/commands/session.rs
@@ -46,7 +46,7 @@ pub fn build_session<'a>(
 
     // TODO: Odd to be prepping the provider rather than having that done in the agent?
     let provider = factory::get_provider(provider_config).unwrap();
-    let agent = AgentFactory::create(agent_version.as_deref().unwrap_or("base"), provider).unwrap();
+    let agent = AgentFactory::create(agent_version.as_deref().unwrap_or("default"), provider).unwrap();
     let prompt = match std::env::var("GOOSE_INPUT") {
         Ok(val) => match val.as_str() {
             "rustyline" => Box::new(RustylinePrompt::new()) as Box<dyn Prompt>,
