+++ crates/goose/src/errors.rs
@@ -19,6 +19,9 @@ pub enum AgentError {
     #[error("Invalid tool name: {0}")]
     InvalidToolName(String),
 
+    #[error("System not found: {0}")]
+    SystemNotFound(String),
+
     #[error("Agent version not found: {0}")]
     VersionNotFound(String),
 }
